@@ -2,12 +2,13 @@
 name = "asset-catalog"
 version = "0.1.0"
 authors = ["mb <mb@nesium.com>"]
-edition = "2018"
+edition = "2021"
 
 [dependencies]
-serde_json = "1.0"
 parser = { path = "../parser" }
+serde_json = "1.0"
+thiserror = { workspace = true }
 
 [dev-dependencies]
-dir-diff = "0.3"
-tempdir = "0.3"
+dir-diff = { workspace = true }
+tempdir = { workspace = true }
