@@ -2,11 +2,12 @@
 name = "swift-gen"
 version = "0.1.0"
 authors = ["mb <mb@nesium.com>"]
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 parser = { path = "../parser" }
+thiserror = { workspace = true }
 
 [dev-dependencies]
-dir-diff = "0.3"
-tempdir = "0.3"
+dir-diff = { workspace = true }
+tempdir = { workspace = true }
