@@ -1,25 +1,29 @@
-use super::ast::{
-  Color, ColorSet, ColorSetValue, Declaration, Document, DocumentItem, RuleSet, RuleSetItem, Value,
-  Variable,
-};
-use super::error::Error;
+use std::fs::File;
+use std::io::prelude::Read;
+use std::num::ParseIntError;
+use std::path::Path;
+use std::str;
+
+use crate::error::Error;
 use nom::branch::alt;
 use nom::bytes::complete::{tag, take_while_m_n};
 use nom::character::complete::{
   alphanumeric1, char, digit1, multispace0, newline, not_line_ending, space0, space1,
 };
 use nom::combinator::{all_consuming, cut, map, map_res, opt};
-use nom::error::{context, convert_error, ParseError, VerboseError};
-use nom::multi::{many0, separated_list};
+use nom::error::{
+  context, convert_error, ContextError, FromExternalError, ParseError, VerboseError,
+};
+use nom::multi::{many0, separated_list0};
 use nom::number::complete::float;
 use nom::sequence::{delimited, preceded, separated_pair, terminated, tuple};
 use nom::Err;
 use nom::IResult;
 
-use std::fs::File;
-use std::io::prelude::Read;
-use std::path::Path;
-use std::str;
+use super::ast::{
+  Color, ColorSet, ColorSetValue, Declaration, Document, DocumentItem, RuleSet, RuleSetItem, Value,
+  Variable,
+};
 
 fn parse_hex_value(input: &str) -> Result<u32, std::num::ParseIntError> {
   u32::from_str_radix(input, 16)
@@ -38,10 +42,12 @@ fn colorset_from_declarations(
       light: decl2.value,
       dark: decl1.value,
     }),
-    _ => Err(Error::new(format!(
-      "Expected light & dark properties. Found {}, {}.",
-      decl1.identifier, decl2.identifier
-    ))),
+    _ => Err(Error::InvalidColorSetDeclaration {
+      message: format!(
+        "Expected light & dark properties. Found {}, {}.",
+        decl1.identifier, decl2.identifier
+      ),
+    }),
   }
 }
 
@@ -52,7 +58,7 @@ pub fn parse_document(input: String) -> Result<Document, Error> {
   let result: IResult<&str, Document, VerboseError<&str>> = map(
     all_consuming(delimited(
       multiline_whitespace,
-      separated_list(
+      separated_list0(
         line_delimiter,
         alt((
           map(variable, DocumentItem::Variable),
@@ -67,9 +73,9 @@ pub fn parse_document(input: String) -> Result<Document, Error> {
 
   match result {
     Ok((_, doc)) => Ok(doc),
-    Err(Err::Error(e)) | Err(Err::Failure(e)) => Err(Error::new(
-      convert_error(&modified_input, e).replace("'\n'", "'\\n'"),
-    )),
+    Err(Err::Error(e)) | Err(Err::Failure(e)) => Err(Error::ParseError {
+      message: convert_error(modified_input.as_ref(), e).replace("'\n'", "'\\n'"),
+    }),
     _ => {
       eprintln!("An unknown error occurred.");
       std::process::exit(0x0100)
@@ -77,35 +83,24 @@ pub fn parse_document(input: String) -> Result<Document, Error> {
   }
 }
 
-pub fn parse_document_from_file(filepath: &str) -> Result<Document, Error> {
-  let path = Path::new(filepath);
-
-  if !path.exists() {
-    return Err(Error::new(format!("No such file {}", filepath)));
-  }
-
-  let mut f = File::open(filepath).unwrap();
+pub fn parse_document_from_file(filepath: impl AsRef<Path>) -> Result<Document, Error> {
+  let mut f = File::open(filepath)?;
   let mut buffer = vec![];
   f.read_to_end(&mut buffer).unwrap();
 
-  let contents = match str::from_utf8(&buffer) {
-    Ok(v) => v,
-    Err(e) => {
-      return Err(Error::new(format!(
-        "Could not read contents of file {}. Reason: {}",
-        filepath, e
-      )))
-    }
-  };
-
+  let contents = str::from_utf8(&buffer)?;
   parse_document(contents.to_string())
 }
 
-fn single_line_comment<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, &str, E> {
+fn single_line_comment<'a, E: ParseError<&'a str> + ContextError<&'a str>>(
+  input: &'a str,
+) -> IResult<&'a str, &str, E> {
   context("Single Line Comment", preceded(tag("//"), not_line_ending))(input)
 }
 
-fn line_delimiter<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, char, E> {
+fn line_delimiter<'a, E: ParseError<&'a str> + ContextError<&'a str>>(
+  input: &'a str,
+) -> IResult<&'a str, char, E> {
   cut(preceded(
     space0,
     preceded(
@@ -115,17 +110,25 @@ fn line_delimiter<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str
   ))(input)
 }
 
-fn multiline_whitespace<'a, E: ParseError<&'a str>>(
+fn multiline_whitespace<'a, E: ParseError<&'a str> + ContextError<&'a str>>(
   input: &'a str,
 ) -> IResult<&'a str, Vec<&str>, E> {
   delimited(
     multispace0,
-    separated_list(terminated(newline, multispace0), single_line_comment),
+    separated_list0(terminated(newline, multispace0), single_line_comment),
     multispace0,
   )(input)
 }
 
-fn ruleset<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, RuleSet, E> {
+fn ruleset<
+  'a,
+  E: ParseError<&'a str>
+    + ContextError<&'a str>
+    + FromExternalError<&'a str, ParseIntError>
+    + FromExternalError<&'a str, Error>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, RuleSet, E> {
   let body = |input| {
     delimited(
       terminated(char('{'), multiline_whitespace),
@@ -149,7 +152,15 @@ fn ruleset<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, RuleS
   )(input)
 }
 
-fn variable<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Declaration<Value>, E> {
+fn variable<
+  'a,
+  E: ParseError<&'a str>
+    + ContextError<&'a str>
+    + FromExternalError<&'a str, ParseIntError>
+    + FromExternalError<&'a str, Error>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, Declaration<Value>, E> {
   context(
     "Variable",
     map(
@@ -166,7 +177,15 @@ fn variable<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Decl
   )(input)
 }
 
-fn value<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Value, E> {
+fn value<
+  'a,
+  E: ParseError<&'a str>
+    + ContextError<&'a str>
+    + FromExternalError<&'a str, ParseIntError>
+    + FromExternalError<&'a str, Error>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, Value, E> {
   preceded(
     space0,
     alt((
@@ -176,7 +195,10 @@ fn value<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Value, 
   )(input)
 }
 
-fn colorset_value<'a, E: ParseError<&'a str>>(
+fn colorset_value<
+  'a,
+  E: ParseError<&'a str> + ContextError<&'a str> + FromExternalError<&'a str, ParseIntError>,
+>(
   input: &'a str,
 ) -> IResult<&'a str, ColorSetValue, E> {
   preceded(
@@ -189,7 +211,12 @@ fn colorset_value<'a, E: ParseError<&'a str>>(
   )(input)
 }
 
-fn rgba_color<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Color, E> {
+fn rgba_color<
+  'a,
+  E: ParseError<&'a str> + ContextError<&'a str> + FromExternalError<&'a str, ParseIntError>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, Color, E> {
   let u8_value = move |input: &'a str| {
     preceded(
       space0,
@@ -225,7 +252,12 @@ fn rgba_color<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Co
   )(input)
 }
 
-fn hex_color<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Color, E> {
+fn hex_color<
+  'a,
+  E: ParseError<&'a str> + ContextError<&'a str> + FromExternalError<&'a str, ParseIntError>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, Color, E> {
   map(
     tuple((
       char('#'),
@@ -243,7 +275,15 @@ fn hex_color<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Col
   )(input)
 }
 
-pub fn colorset<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, ColorSet, E> {
+pub fn colorset<
+  'a,
+  E: ParseError<&'a str>
+    + ContextError<&'a str>
+    + FromExternalError<&'a str, ParseIntError>
+    + FromExternalError<&'a str, Error>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, ColorSet, E> {
   context(
     "ColorSet",
     map_res(
@@ -261,7 +301,9 @@ pub fn colorset<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, 
   )(input)
 }
 
-pub fn identifier<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, String, E> {
+pub fn identifier<'a, E: ParseError<&'a str> + ContextError<&'a str>>(
+  input: &'a str,
+) -> IResult<&'a str, String, E> {
   context(
     "Identifier",
     map(preceded(space0, alphanumeric1), |ident: &str| {
@@ -270,7 +312,7 @@ pub fn identifier<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str
   )(input)
 }
 
-pub fn variable_identifier<'a, E: ParseError<&'a str>>(
+pub fn variable_identifier<'a, E: ParseError<&'a str> + ContextError<&'a str>>(
   input: &'a str,
 ) -> IResult<&'a str, String, E> {
   context(
@@ -282,7 +324,12 @@ pub fn variable_identifier<'a, E: ParseError<&'a str>>(
   )(input)
 }
 
-pub fn variable_value<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, Variable, E> {
+pub fn variable_value<
+  'a,
+  E: ParseError<&'a str> + ContextError<&'a str> + FromExternalError<&'a str, ParseIntError>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, Variable, E> {
   map(
     tuple((variable_identifier, opt(preceded(space1, alpha_value)))),
     |res| {
@@ -295,7 +342,7 @@ pub fn variable_value<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a
   )(input)
 }
 
-pub fn declaration<'a, F, V, E: ParseError<&'a str>>(
+pub fn declaration<'a, F, V, E: ParseError<&'a str> + ContextError<&'a str>>(
   value: F,
 ) -> impl Fn(&'a str) -> IResult<&'a str, Declaration<V>, E>
 where
@@ -315,14 +362,24 @@ where
   }
 }
 
-fn hex_value<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, u32, E> {
+fn hex_value<
+  'a,
+  E: ParseError<&'a str> + ContextError<&'a str> + FromExternalError<&'a str, ParseIntError>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, u32, E> {
   context(
     "Hex Value",
     map_res(take_while_m_n(6, 6, is_hex_digit), parse_hex_value),
   )(input)
 }
 
-fn alpha_value<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, f32, E> {
+fn alpha_value<
+  'a,
+  E: ParseError<&'a str> + ContextError<&'a str> + FromExternalError<&'a str, ParseIntError>,
+>(
+  input: &'a str,
+) -> IResult<&'a str, f32, E> {
   context(
     "Alpha Value",
     map_res(terminated(digit1, tag("%")), |input| {
