@@ -1,8 +1,8 @@
 mod colorset_renderer;
-mod dynamic_color_renderer;
 pub mod data;
+mod dynamic_color_renderer;
 mod renderer;
 
-pub use renderer::{Renderer, RendererConfig};
 pub use colorset_renderer::ColorSetRenderer;
 pub use dynamic_color_renderer::DynamicColorRenderer;
+pub use renderer::{Renderer, RendererConfig};
